data_processing/tools/extract_events_from_document/pyproject.toml
[project]
name = "extract-events-from-document"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Cyril Mammar", email = "mammar.cyril@epitech.eu" }
]
requires-python = ">=3.12"
dependencies = [
    "dnspython>=2.7.0",
    "google-adk==1.1.1",
    "pymongo>=4.13.0",
    "pytest>=8.3.5",
    "tinydb>=4.8.2",
    "common"
]

[project.scripts]
extract-events-from-document = "extract_events_from_document.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

#dis a uv d'aller chercher common en local par sur pypi
[tool.uv.sources]
common = { workspace = true }

[tool.hatch.build.targets.wheel]
packages = ["src/extract_events_from_document"]